refactor(game): simplify Java integrity flow
